Fix duplicated lightning and ineffective city-mode energy metering
- Lightning rod: drop the redundant spawnEntity call. addWeatherEffect
  already registers and syncs the bolt, so spawning it again double-ticked
  its damage, ignition, and effects.
- Energy meter: cast the city-mode passthrough amount to double so it
  binds to the overload the meter actually overrides. The int overload is
  a no-op default, so the meter previously recorded nothing in city mode.

+++ src/main/java/blusunrize/immersiveengineering/common/blocks/metal/TileEntityConnectorLV.java
@@ -385,10 +385,10 @@ public class TileEntityConnectorLV extends TileEntityImmersiveConnectable implem
 				{
 					IImmersiveConnectable subStart = ApiUtils.toIIC(sub.start, world);
 					if(subStart!=null&&passed.add(subStart))
-						subStart.onEnergyPassthrough(sent);
+						subStart.onEnergyPassthrough((double)sent);
 					IImmersiveConnectable subEnd = ApiUtils.toIIC(sub.end, world);
 					if(subEnd!=null&&passed.add(subEnd))
-						subEnd.onEnergyPassthrough(sent);
+						subEnd.onEnergyPassthrough((double)sent);
 				}
 			}
 		}
